+++ admin/upload.php
@@ -13,12 +13,19 @@
 $load['plugin'] = true;
 include('inc/common.php');
 login_cookie_check();
+$dirsSorted=null;$filesSorted=null;$foldercount=null;
 
+if (isset($_GET['path'])) {
+	$path = str_replace('../','', $_GET['path']);
+	$path = tsl("../data/uploads/".$path);
+	$subPath = str_replace('../','', $_GET['path']);
+	$subFolder = tsl($subPath);
+} else { 
+	$path = "../data/uploads/";
+	$subPath = ''; 
+	$subFolder = '';
+}
 
-$path = (isset($_GET['path'])) ? "../data/uploads/".$_GET['path'] : "../data/uploads/";
-$subPath = (isset($_GET['path'])) ? $_GET['path'] : "";
-$subFolder = (empty($subPath)) ? "" : $subPath."/";
-$path = tsl($path);
 // check if host uses Linux (used for displaying permissions
 $isUnixHost = (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN' ? false : true);
 
@@ -45,14 +52,22 @@ if (isset($_FILES['file'])) {
 				$count++;
 			}
 			
-			//create file
-			move_uploaded_file($_FILES["file"]["tmp_name"][$i], $file_loc);
-			
-			//run file upload hook
-			exec_action('file-uploaded');
+			//validate file
+			if (validate_safe_file($_FILES["file"]["tmp_name"][$i], $_FILES["file"]["name"][$i],  $_FILES["file"]["type"][$i])) {
+				move_uploaded_file($_FILES["file"]["tmp_name"][$i], $file_loc);
+				if (defined('GSCHMOD')) {
+					chmod($file_loc, GSCHMOD);
+				} else {
+					chmod($file_loc, 0644);
+				}
+				exec_action('file-uploaded');
+				$messages[] = i18n_r('FILE_SUCCESS_MSG').': <a href="'. $SITEURL .'data/uploads/'.$subFolder.$base.'">'. $SITEURL .'data/uploads/'.$subFolder.$base.'</a>';
+			} else {
+				$messages[] = $_FILES["file"]["name"][$i] .' - '.i18n_r('ERROR_UPLOAD');
+			}
 			
 			//successfull message
-			$messages[] = i18n_r('FILE_SUCCESS_MSG').': <a href="'. $SITEURL .'data/uploads/'.$subFolder.$base.'">'. $SITEURL .'data/uploads/'.$subFolder.$base.'</a>';
+			
 		}
 	 }
 	 // after uploading all files process messages
@@ -70,6 +85,15 @@ if (isset($_FILES['file'])) {
 }
 // if creating new folder
 if (isset($_GET['newfolder'])) {
+	
+	// check for csrf
+	if (!defined('GSNOCSRF') || (GSNOCSRF == FALSE) ) {
+		$nonce = $_GET['nonce'];
+		if(!check_nonce($nonce, "createfolder")) {
+			die("CSRF detected!");
+		}
+	}
+	
 	$newfolder = $_GET['newfolder'];
 	// check for invalid chars
 	$cleanname = clean_url(to7bit($newfolder, "UTF-8"));
@@ -91,13 +115,14 @@ if (isset($_GET['newfolder'])) {
 		}
 	}
 }
-?>
-<?php get_template('header', cl($SITENAME).' &raquo; '.i18n_r('FILE_MANAGEMENT')); ?>
 
-	<h1><a href="<?php echo $SITEURL; ?>" target="_blank" ><?php echo cl($SITENAME); ?></a> <span>&raquo;</span> <?php echo i18n_r('FILE_MANAGEMENT'); ?></h1>
-	<?php include('template/include-nav.php');?>
-	<?php include('template/error_checking.php');?>
-	<div class="bodycontent">
+get_template('header', cl($SITENAME).' &raquo; '.i18n_r('FILE_MANAGEMENT')); 
+
+?>
+	
+<?php include('template/include-nav.php'); ?>
+
+<div class="bodycontent clearfix">
 	<div id="maincontent">
 		<div class="main" >
 		<h3 class="floated"><?php echo i18n('UPLOADED_FILES'); ?><span id="filetypetoggle">&nbsp;&nbsp;/&nbsp;&nbsp;<?php echo i18n('SHOW_ALL'); ?></span></h3>
@@ -117,6 +142,9 @@ if (isset($_GET['newfolder'])) {
             // not a upload file
           	} elseif (is_dir($path . $file)) {
             $dirsArray[$dircount]['name'] = $file;
+            clearstatcache();
+						$ss = @stat($path . $file);
+						$dirsArray[$dircount]['date'] = @date('M j, Y',$ss['ctime']);
             $dircount++;
 					} else {
 						$filesArray[$count]['name'] = $file;
@@ -176,38 +204,50 @@ if (isset($_GET['newfolder'])) {
       }
       echo '</div> <div id="new-folder">
       	<a href="#" id="createfolder">'.i18n_r('CREATE_FOLDER').'</a>
-				<form action="upload.php">&nbsp;<input type="hidden" name="path" value="'.$subPath.'" /><input type="text" class="text" name="newfolder" id="foldername" /> <input type="submit" class="submit" value="'.i18n_r('CREATE_FOLDER').'" />&nbsp; <a href="#" class="cancel">'.i18n_r('CANCEL').'</a></form>
+				<form action="upload.php">&nbsp;<input type="hidden" name="path" value="'.$subPath.'" /><input type="hidden" name="nonce" value="'. get_nonce("createfolder") .'" /><input type="text" class="text" name="newfolder" id="foldername" /> <input type="submit" class="submit" value="'.i18n_r('CREATE_FOLDER').'" />&nbsp; <a href="#" class="cancel">'.i18n_r('CANCEL').'</a></form>
 			</div></div>';
       
 			
 			
      echo '<table class="highlight" id="imageTable">'; 
-     echo '<tr><th class="imgthumb" ></th><th>'.i18n_r('FILE_NAME').'</th><th style="text-align:right;">'.i18n_r('FILE_SIZE').'</th>';
+     echo '<tr><th class="imgthumb" ></th><th>'.i18n_r('FILE_NAME').'</th>';
+     echo '<th style="text-align:right;">'.i18n_r('FILE_SIZE').'</th>';
+     echo '<th style="text-align:right;">'.i18n_r('DATE').'</th>';
      if (defined('GSDEBUG')){
-     	echo '<th style="text-align:right;">'.i18n_r('PERMS').'</th>';
-		$cols='4';
-     } else {
-     	$cols='3';
+     	 echo '<th style="text-align:right;">'.i18n_r('PERMS').'</th>';
      }
-     echo '<th style="text-align:right;">'.i18n_r('DATE').'</th><th></th></tr>';  
+     echo '<th><!-- actions --></th></tr>';  
      if (count($dirsSorted) != 0) {
+     		$foldercount = 0;
         foreach ($dirsSorted as $upload) {
+        	
+        	$upload['name'] = rawurlencode($upload['name']);
         	
         	# check to see if folder is empty
         	$directory_delete = null;
         	if ( check_empty_folder($path.$upload['name']) ) {  
-						$directory_delete = '<a class="delconfirm" title="'.i18n_r('DELETE_FOLDER').': '. $upload['name'] .'" href="deletefile.php?path='.$urlPath.'&amp;folder='. $upload['name'] . '&amp;nonce='.get_nonce("delete", "deletefile.php").'">X</a>';
+						$directory_delete = '<a class="delconfirm" title="'.i18n_r('DELETE_FOLDER').': '. $upload['name'] .'" href="deletefile.php?path='.$urlPath.'&amp;folder='. $upload['name'] . '&amp;nonce='.get_nonce("delete", "deletefile.php").'">&times;</a>';
 					}
+        	$directory_size = '<span>'.folder_items($path.$upload['name']).' '.i18n_r('ITEMS').'</span>';
         	
-          echo '<tr class="All folder" >';
-          echo '<td class="imgthumb" ></td><td colspan="'.$cols.'">';
+          echo '<tr class="All folder '.$upload['name'].'" >';
+          echo '<td class="imgthumb" ></td><td>';
         
           $adm = substr($path . $upload['name'] ,  16); 
-          echo '<img src="template/images/folder.png" width="11" /> <a href="upload.php?path='.$adm.'" ><strong>'.$upload['name'].'</strong></a>';
-                   
-          echo '</td>';
+          echo '<img src="template/images/folder.png" width="11" /> <a href="upload.php?path='.$adm.'" ><strong>'.$upload['name'].'</strong></a></td>';
+          echo '<td style="width:80px;text-align:right;" ><span>'.$directory_size.'</span></td>';
+          
+          // get the file permissions.
+					if ($isUnixHost && defined('GSDEBUG') && function_exists('posix_getpwuid')) {
+						$filePerms = substr(sprintf('%o', fileperms($path.$upload['name'])), -4);
+						$fileOwner = posix_getpwuid(fileowner($path.$upload['name']));
+						echo '<td style="width:70px;text-align:right;"><span>'.$fileOwner['name'].'/'.$filePerms.'</span></td>';
+					}
+					
+		      echo '<td style="width:85px;text-align:right;" ><span>'. shtDate($upload['date']) .'</span></td>';
           echo '<td class="delete" >'.$directory_delete.'</td>';
           echo '</tr>';
+          $foldercount++;
         }
      }
 			if (count($filesSorted) != 0) { 			
@@ -221,15 +261,15 @@ if (isset($_GET['newfolder'])) {
 					echo '<tr class="All '.$upload['type'].' '.$cclass.'" >';
 					echo '<td class="imgthumb" >';
 					if ($upload['type'] == i18n_r('IMAGES') .' Images') {
-						$gallery = 'rel="facybox"';
+						$gallery = 'rel=" facybox_i"';
 						$pathlink = 'image.php?i='.$upload['name'].'&amp;path='.$subPath;
 						$thumbLink = $urlPath.'thumbsm.'.$upload['name'];
-						if (file_exists('../data/thumbs/'.$thumbLink)) {
+						if (file_exists(GSTHUMBNAILPATH.$thumbLink)) {
 							$imgSrc='<img src="../data/thumbs/'. $thumbLink .'" />';
 						} else {
-							$imgSrc='<img src="inc/thumb.php?src='. $urlPath . $upload['name'] .'&amp;dest='. $thumbLink .'&amp;x=65&amp;f=1" />';
+							$imgSrc='<img src="inc/thumb.php?src='. $urlPath . $upload['name'] .'&amp;dest='. $thumbLink .'&amp;f=1" />';
 						}
-						echo '<a href="'. $path . $upload['name'] .'" title="'. $upload['name'] .'" rel="facybox" >'.$imgSrc.'</a>';
+						echo '<a href="'. $path . $upload['name'] .'" title="'. $upload['name'] .'" rel=" facybox_i" >'.$imgSrc.'</a>';
 					} else {
 						$gallery = '';
 						$controlpanel = '';
@@ -239,32 +279,35 @@ if (isset($_GET['newfolder'])) {
 					echo '<td style="width:80px;text-align:right;" ><span>'. $upload['size'] .'</span></td>';
              
 		            
-					if ($isUnixHost && defined('GSDEBUG')) {
+					// get the file permissions.
+					if ($isUnixHost && defined('GSDEBUG') && function_exists('posix_getpwuid')) {
 						$filePerms = substr(sprintf('%o', fileperms($path.$upload['name'])), -4);
-						$fileOwner = null;
-						if (function_exists('posix_getpwuid')) {
-							$fileOwner = posix_getpwuid(fileowner($path.$upload['name']));
-						}
-						if (($filePerms) && ($fileOwner['name'])){
-							echo '<td style="width:70px;text-align:right;"><span>'.$fileOwner['name'].'/'.$filePerms.'</span></td>';
-						}
+						$fileOwner = posix_getpwuid(fileowner($path.$upload['name']));
+						echo '<td style="width:70px;text-align:right;"><span>'.$fileOwner['name'].'/'.$filePerms.'</span></td>';
 					}
-					
+							
 					echo '<td style="width:85px;text-align:right;" ><span>'. shtDate($upload['date']) .'</span></td>';
-					echo '<td class="delete" ><a class="delconfirm" title="'.i18n_r('DELETE_FILE').': '. htmlspecialchars($upload['name']) .'" href="deletefile.php?file='. $upload['name'] . '&amp;path=' . $urlPath . '&amp;nonce='.get_nonce("delete", "deletefile.php").'">X</a></td>';
+					echo '<td class="delete" ><a class="delconfirm" title="'.i18n_r('DELETE_FILE').': '. htmlspecialchars($upload['name']) .'" href="deletefile.php?file='. $upload['name'] . '&amp;path=' . $urlPath . '&amp;nonce='.get_nonce("delete", "deletefile.php").'">&times;</a></td>';
 					echo '</tr>';
 					exec_action('file-extras');
 				}
 			}
 			echo '</table>';
-			echo '<p><em><b>'. $counter .'</b> '.i18n_r('TOTAL_FILES').' ('. fSize($totalsize) .')</em></p>';
+			if ($counter > 0) { 
+				$sizedesc = '('. fSize($totalsize) .')';
+			} else {
+				$sizedesc = '';
+			}
+			$totalcount = (int)$counter+(int)$foldercount;
+			echo '<p><em><b><span id="pg_counter">'. $totalcount .'</span></b> '.i18n_r('TOTAL_FILES').' '.$sizedesc.'</em></p>';
 		?>	
 		</div>
 		</div>
 	</div>
-		<div id="sidebar" >
-		<?php include('template/sidebar-files.php'); ?>
-		</div>	
-	<div class="clear"></div>
-	</div>
+	
+	<div id="sidebar" >
+	<?php include('template/sidebar-files.php'); ?>
+	</div>	
+	
+</div>
 <?php get_template('footer'); ?>
